correct type declartion in ParserBase (allow numbers to be template paramaters)

+++ include/TFEL/Math/tmatrix.hxx
@@ -126,7 +126,7 @@ namespace tfel{
        * type of the size of the container.
        * (this is a stl requirement).
        */
-      typedef size_t size_type;
+      typedef unsigned short size_type;
       /*!
        * type of the difference between two iterators.
        * (this is a stl requirement).
